+++ src/library/PostgREST/Error.hs
@@ -83,7 +83,12 @@ errorResponseFor verb err =
   let
     baseHeader = MediaType.toContentType MTApplicationJSON
     cLHeader body = (,) "Content-Length" (show $ LBS.length body) :: Header
-    pSHeader code' = ("Proxy-Status", "PostgREST; error=" <> T.encodeUtf8 code')
+    pSHeader code' =
+      ("Proxy-Status", "PostgREST; error=" <> T.encodeUtf8 code' <> detailsHeader)
+      where
+        detailsHeader = if code' == "PGRST303"
+          then maybe mempty (\d -> "; details=" <> LBS.toStrict (JSON.encode d)) (details err)
+          else mempty
   in
   responseLBS (status err) (baseHeader : cLHeader (errorPayload verb err) : pSHeader (code err) : headers err) $ errorPayload verb err