+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file. From versio
 ### Changes
 
 - Error response body now includes details for JWT time validation errors by @steve-chavez in #5196
+- `Proxy-Status` now includes a `details` parameter for `PGRST303` error codes by @steve-chavez in #5196
 
 ## [16.2] - 2026-08-21